ci: label rebuild counts from report publishing
