+++ tools/wasinix/src/cli/mod.rs
@@ -501,13 +501,6 @@ pub enum CiCommand {
         #[arg(long)]
         out_dir: PathBuf,
     },
-    /// Reconcile a managed update PR from a completed CI diff report
-    ReconcileUpdatePullRequest {
-        #[arg(long)]
-        run_dir: PathBuf,
-        #[command(flatten)]
-        surface: crate::github::surfaces::SurfaceArgs,
-    },
     /// Record a workflow run's step durations: a step-summary table, and
     /// with --publish a document the timings fold reads back
     StepTimings {
@@ -1673,6 +1666,7 @@ fn ci_command(command: CiCommand) -> Result<CommandStatus> {
                 return Ok(CommandStatus::SUCCESS);
             }
             let rendered = crate::github::publish::load(&run_dir)?;
+            crate::github::publish::reconcile_rebuild_label(&client, &rendered, &target, effects)?;
             if failure_logs {
                 let sha = target.head_sha.clone().ok_or_else(|| {
                     crate::support::error::Error::Request(
@@ -1842,58 +1836,6 @@ fn ci_command(command: CiCommand) -> Result<CommandStatus> {
         }
         CiCommand::MutatePublish { out_dir } => {
             crate::github::mutation::mutate_publish(&repo, &out_dir)?;
-            Ok(CommandStatus::SUCCESS)
-        }
-        CiCommand::ReconcileUpdatePullRequest { run_dir, surface } => {
-            let pull_request = surface.pull_request.ok_or_else(|| {
-                crate::support::error::Error::Request(
-                    "update PR reconciliation needs --pull-request".into(),
-                )
-            })?;
-            let repository = surface.repository(&repo)?;
-            let client =
-                crate::github::client::Client::new(crate::github::client::token().as_deref());
-            let pull = crate::github::mutation::pull(&client, &repository, pull_request)?;
-            let Some(state) = crate::update::managed::decode(&pull.body)? else {
-                ui::note("not a managed update pull request");
-                return Ok(CommandStatus::SUCCESS);
-            };
-            if !matches!(
-                crate::update::managed::parse_recipe(&state.recipe)?,
-                CommandTree::Update(_)
-            ) {
-                ui::note("managed pull request is not an update");
-                return Ok(CommandStatus::SUCCESS);
-            }
-            let report: crate::ci::report::Report =
-                schema::read(&crate::ci::prepare::report_path(&run_dir))?;
-            crate::support::error::require(report.complete, "CI report is incomplete")?;
-            let comparison = match report.comparisons.as_slice() {
-                [comparison] if comparison.base_evaluated && comparison.head_evaluated => {
-                    comparison
-                }
-                _ => {
-                    return crate::support::error::request_error(
-                        "CI report has no complete diff comparison",
-                    )
-                }
-            };
-            let eval = comparison.eval.as_ref().ok_or_else(|| {
-                crate::support::error::Error::Request("CI report has no evaluated diff".into())
-            })?;
-            let changed = eval
-                .added
-                .iter()
-                .chain(&eval.removed)
-                .chain(&eval.rebuilt)
-                .collect::<std::collections::BTreeSet<_>>();
-            crate::github::update_pr::reconcile_rebuild_label(
-                &client,
-                &repository,
-                pull_request,
-                changed.len(),
-            )?;
-            ui::fact("rebuild count", changed.len());
             Ok(CommandStatus::SUCCESS)
         }
         CiCommand::StepTimings {
